+++ src/test/groovy/wslite/http/HTTPClientSpec.groovy
@@ -31,6 +31,8 @@ class HTTPClientSpec extends Specification {
         httpc.httpConnectionFactory = getHTTPConnectionFactory(conn)
         httpc.connectTimeout = 20000
         httpc.readTimeout = 15000
+        httpc.followRedirects = false
+        httpc.useCaches = false
 
         when:
         def request = new HTTPRequest(url:testURL, method:HTTPMethod.GET)
@@ -39,6 +41,8 @@ class HTTPClientSpec extends Specification {
         then:
         conn.connectTimeout == 20000
         conn.readTimeout == 15000
+        !conn.instanceFollowRedirects
+        !conn.useCaches
     }
 
     def "request settings will override settings"() {
@@ -46,16 +50,103 @@ class HTTPClientSpec extends Specification {
         httpc.httpConnectionFactory = getHTTPConnectionFactory(conn)
         httpc.connectTimeout = 20000
         httpc.readTimeout = 15000
+        httpc.followRedirects = false
+        httpc.useCaches = true
 
         when:
         def request = new HTTPRequest(url:testURL, method:HTTPMethod.GET)
         request.connectTimeout = 7000
         request.readTimeout = 9000
+        request.followRedirects = true
+        request.useCaches = false
         httpc.execute(request)
 
         then:
         conn.connectTimeout == 7000
         conn.readTimeout == 9000
+        conn.instanceFollowRedirects
+        !conn.useCaches
+    }
+
+    def "client will trust all ssl certs if not overridden on request"() {
+        given:
+        httpc.httpConnectionFactory = getHTTPConnectionFactory(conn)
+        httpc.sslTrustAllCerts = true
+
+        when:
+        def request = new HTTPRequest(url:"https://foo.org".toURL(), method:HTTPMethod.GET)
+        httpc.execute(request)
+
+        then:
+        conn.hostnameVerifier.verify("foo", null)
+    }
+
+    def "request set to trust all ssl certs will override client"() {
+        given:
+        httpc.httpConnectionFactory = getHTTPConnectionFactory(conn)
+        httpc.sslTrustAllCerts = false
+
+        when:
+        def request = new HTTPRequest(url:"https://foo.org".toURL(), method:HTTPMethod.GET)
+        request.sslTrustAllCerts = true
+        httpc.execute(request)
+
+        then:
+        conn.hostnameVerifier.verify("foo", null)
+    }
+
+    def "request set to not trust all ssl will override client"() {
+        given:
+        httpc.httpConnectionFactory = getHTTPConnectionFactory(conn)
+        httpc.sslTrustAllCerts = true
+
+        when:
+        def request = new HTTPRequest(url:"https://foo.org".toURL(), method:HTTPMethod.GET)
+        request.sslTrustAllCerts = false
+        httpc.execute(request)
+
+        then:
+        null == conn.hostnameVerifier
+    }
+
+    def "client will use trust store if not overridden on request"() {
+        given:
+        httpc.httpConnectionFactory = [getConnectionUsingTrustStore:{url, tsfile, tspassword, proxy=null ->
+            conn.URL = url
+            conn.setRequestProperty("javax.net.ssl.trustStore", tsfile)
+            conn.setRequestProperty("javax.net.ssl.trustStorePassword", tspassword)
+            return conn
+        }] as HTTPConnectionFactory
+        httpc.sslTrustStoreFile = "~/test.jks"
+
+        when:
+        def request = new HTTPRequest(url:"https://foo.org".toURL(), method:HTTPMethod.GET)
+        httpc.execute(request)
+
+        then:
+        "~/test.jks" == conn.requestProperties["javax.net.ssl.trustStore"]
+        null == conn.requestProperties["javax.net.ssl.trustStorePassword"]
+    }
+
+    def "request trust store settings will override client"() {
+        given:
+        httpc.httpConnectionFactory = [getConnectionUsingTrustStore:{url, tsfile, tspassword, proxy=null ->
+            conn.URL = url
+            conn.setRequestProperty("javax.net.ssl.trustStore", tsfile)
+            conn.setRequestProperty("javax.net.ssl.trustStorePassword", tspassword)
+            return conn
+        }] as HTTPConnectionFactory
+        httpc.sslTrustStoreFile = "~/test.jks"
+        httpc.sslTrustStorePassword = "foo"
+
+        when:
+        def request = new HTTPRequest(url:"https://foo.org".toURL(), method:HTTPMethod.GET)
+        request.sslTrustStoreFile = "~/usethis.jks"
+        httpc.execute(request)
+
+        then:
+        "~/usethis.jks" == conn.requestProperties["javax.net.ssl.trustStore"]
+        null == conn.requestProperties["javax.net.ssl.trustStorePassword"]
     }
 
 }