+++ src/test/groovy/wslite/rest/RequestBuilderSpec.groovy
@@ -108,20 +108,20 @@ class RequestBuilderSpec extends Specification {
     def "sets http connection parameters"() {
         given:
         def url =  "http://ws.org/services"
-        def params = [path: "/foo", readTimeout:9876, trustAllSSLCerts:false]
+        def params = [path: "/foo", readTimeout:9876, sslTrustAllCerts:false]
 
         when:
         HTTPRequest request = new RequestBuilder().build(HTTPMethod.GET, url, params, null)
 
         then:
         9876 == request.readTimeout
-        !request.trustAllSSLCerts
+        !request.sslTrustAllCerts
     }
 
     def "original params not modified"() {
         given:
         def url =  "http://ws.org/services"
-        def params = [path: "/users/123", readTimeout:9876, trustAllSSLCerts:false]
+        def params = [path: "/users/123", readTimeout:9876, sslTrustAllCerts:false]
 
         when:
         HTTPRequest request = new RequestBuilder().build(HTTPMethod.GET, url, params, null)